+++ scripts/new-project.ps1
@@ -1,3 +1,4 @@
+$OutputEncoding = [Console]::OutputEncoding = [System.Text.Encoding]::UTF8;
 [CmdletBinding()]
 param(
     [Parameter(Mandatory=$true)]